Network service refactoring
